Crate ethers_contract_derive

Procedural macros for generating type-safe bindings to an Ethereum smart contract.

Macros

abigen

Generates type-safe bindings to an Ethereum smart contract from its ABI.

Derive Macros

Eip712

Derives the Eip712 trait for the labeled type.

EthAbiCodec

Derives the AbiEncode and AbiDecode traits for the labeled type.

EthAbiType

Derives the AbiType and all Tokenizable traits for the labeled type.

EthCall

Derives the EthCall and Tokenizeable trait for the labeled type.

EthDisplay

Derives the Display trait on structs by formatting each field based on its Ethereum type.

EthError

Derives the EthError and Tokenizeable trait for the labeled type.

EthEvent

Derives the EthEvent and Tokenizable traits for the labeled type.